requirements.txt
openai==0.4.0
moviepy.editor==1.0.3